requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
kiwisolver==1.4.5
packaging==24.0
pillow==10.3.0
pycurl==7.45.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
urllib3==2.2.1
